now the project has a .env archive that holds global variables and the logic of the program is divided in parts
