+++ utn/ignacio_martinez/practices/mongo practice/.gitignore
@@ -1,1 +1,2 @@
 /node_modules
+.env
