+++ utn/ignacio_martinez/practices/mongo practice/config/config.js
@@ -1,6 +1,6 @@
 module.exports = {
     mongo: {
-        MONGO_URI:"mongodb://ignacio_martinez:ignacio_martinez@localhost:27017"
+        MONGO_URI:process.env.MONGO_URI
     },
     mysql: {
