+++ utn/ignacio_martinez/practices/mongo practice/routes/indexRoute.js
@@ -1,0 +1,10 @@
+const express = require("express");
+const route = express.Router();
+const {handleWelcome} = require("../controllers/welcomeController");
+const route_users = require("./userRoute");
+
+route.get("/",handleWelcome);
+
+route.use("/users", route_users);
+
+module.exports = route;
